nvmf/tcp: Fix idle/busy poller status
spdk_sock_group_poll returns the number of processed events,
if it is >0 then the poller was busy. But `rc` variable
is later might be overwritten with result of nvmf_tcp_sock_process
which typically returns 0. That leads to incorrect statistics -
tcp poller reports it was idle most of time.

Signed-off-by: Alexey Marchuk <alexeymar@nvidia.com>
Change-Id: I26b34c1d1ea87afd48b9b3f81d35edc590dc7e29
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/23529
Reviewed-by: Ben Walker <ben@nvidia.com>
Community-CI: Mellanox Build Bot
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>