scheduler/dynamic: adjust frequency when core goes to sleep or wakes
When all spdk_threads have been moved to the scheduling core because
they are all idle, we will try to reduce frequency on the scheduling
core if it is still idle. But the frequency changes don't take effect
if it has an SMT sibling which is still set to max frequency.

So when we have moved the last thread off of a core, set its frequency
to min. And conversely, when we move the first thread onto a core,
set its frequency to max.

Signed-off-by: Jim Harris <jim.harris@samsung.com>
Change-Id: I034194d553aa678d387b1b5448631ba4ffcbb3d3
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/23649
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>