nvmf/auth: add method for changing keys
This allows users to change the DH-HMAC-CHAP keys without
removing/re-adding hosts to a subsystem.  That means that keys can be
changed without terminating existing connections.

Once the keys are changed, new connections, as well as reauthentication
requests on the already established connections, will be required to use
the new keys.

Change-Id: I09e112ae73e4afebd452a3ebd19264cc8217d1f8
Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/23822
Reviewed-by: Krzysztof Karas <krzysztof.karas@intel.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Ben Walker <ben@nvidia.com>
Community-CI: Mellanox Build Bot