nvme: check hostnqn when attaching controllers
This ensures that a new controller will be attached when a controller
with the same transport id, but a different hostnqn already exists.  It
can be useful in cases where the target returns different namespace
configuration based on the NQN of the host.

Additionally, because hostnqns are now always checked and the default
hostnqn is built from nvme_driver.default_extended_host_id, unit tests
had to be changed to ensure that this field is initialized.

Suggested-by: Jim Harris <jim.harris@samsung.com>
Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: Ibb660d7f99ad67ff7c29658a7a7da7d8328b19d6
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/23718
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Reviewed-by: Krzysztof Karas <krzysztof.karas@intel.com>
Community-CI: Mellanox Build Bot