scheduler/gscheduler: decide freq on highest util of SMT siblings
The CPU frequency governor can only take effect if all SMT siblings
agree on the same frequency. Assume there are two SMT siblings,
one is 80% busy, the other is 20% busy, we want to make the decision
on frequency for both cores based on the 80% busy. This ensures the
work on that 80% core is not impacted by the other core trying to set
20%.

Signed-off-by: Jim Harris <jim.harris@samsung.com>
Change-Id: Ifffffaa90dd04285467235b55a3009c21fe45d57
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/23724
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>