nvme: add lock_depth for ctrlr_lock
This supplements checking pthread_mutex_destroy() return code for
detecting if the ctrlr lock is destroy while it is held. At least
on some systems, pthread_mutex_destroy() returns success even if
the lock is currently held.

Note: this is part of patch series that reproduced issue #3401,
and this patch reliably catches that the lock is held when
destroyed after a controller loss timeout from the bdev nvme module.

Signed-off-by: Jim Harris <jim.harris@samsung.com>
Change-Id: Id2a4e3dd9f5d52b3038143ff9c0dceb779b55b56
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/23735
Community-CI: Mellanox Build Bot
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>