rpc.py: use user-specified timeout when executing script
Generally this was working, but bdev_lvol_create_lvstore recently
changed to set its own default (90 seconds) if the user had not
passed one.

The original script execution code didn't account for individual
RPCs doing this, which would result in this RPC and those following
in a script to take on the 90 second default.

Fixes: 679c3183 ("lvol: set default timeout to 90.0 in bdev_lvol_create_lvstore")

Signed-off-by: Jim Harris <jim.harris@samsung.com>
Change-Id: I2d5b88d74ff621392e06040609a305a8aaebcb6c
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/23766
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-by: Changpeng Liu <changpeng.liu@intel.com>
Community-CI: Mellanox Build Bot