bdevperf/bdevperf.py: extend RPC timeout if needed
This patch is created to automatically extend
RPC timeout, which by default is set to 60 sec.,
if someone starts a test with longer duration.
E.g. the call like this:
bdevperf.py perform_tests -q 16 -o 4096 -w write -t 70
would normally result in RPC timeout during waiting for
the response, because the scheduled test will take ~70 seconds,
while the RPC timeout is 60. With this change, RPC timeout
will be set to 75 sec (5 sec longer than expected test), to
allow ample time for the test to complete.

Also, if the user explicitly specifies RPC timeout that is
shorter than the test duration, it will also be extended to
allow time for the test to complete.
I.e., for the call:
bdevperf.py -t 10 perform_tests -q 16 -o 4096 -w write -t 30
the RPC timeout will be automatically extended to 35 sec,
even though the user specified 10 sec (which is insufficient
in this case)

Change-Id: I094f2d2e27b1d363f81977394d6c2fe8f39773eb
Signed-off-by: Marcin Spiewak <marcin.spiewak@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/23776
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Community-CI: Mellanox Build Bot
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>