test/lvol: Use NOT() for negative tests
In these examples, it's expected from the rpc to fail so NOT()
should be used to indicate that.

The actual pipeline variant causes the ERR trap to be triggered
from rpc_cmd() since the PIPESTATUS[@] == 1 0. Due to that, extra
call to print_backtrace() is made from rpc_cmd() context - this
change will also make sure this extra call is avoided.

Change-Id: If445ab4805e53b7d02d2287f74ab061600907d93
Signed-off-by: Michal Berger <michal.berger@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/23793
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Kamil Godzwon <kamilx.godzwon@intel.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>