lib/thread: Align spdk_thread allocation on cache line
spdk_thread is a special object in SPDK because it is not allocated on
the core that is going to use it.
Consecutive mallocs on the same core may share the same cache line. So
it is possible that a spdk_thread object shares a cache line with any
other object. The problem is that the spdk_thread object will be used by
the core actually running the thread and the 2nd object will likely be
used by the core that created the spdk_thread. This is bad for
performance and it is even worse if the two cores are not in the same
socket.

Change-Id: I0628b64ee261901925b7005ba8daee235fdbd592
Signed-off-by: Alex Michon <amichon@kalrayinc.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/23814
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>
Community-CI: Mellanox Build Bot
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>