thread: document that spdk_poller_unregister() can take pointer to NULL
We encourage taking advantage of this in code reviews, but the
header file doesn't say anything about it which makes it difficult
for new users of the API to know it upfront.

While here, also document that this function will write NULL to
the ppoller output parameter and what that is useful.

Signed-off-by: Jim Harris <jim.harris@samsung.com>
Change-Id: Iad06cbb48ddd01b14625257937b289f2f34d53ad
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/23933
Community-CI: Mellanox Build Bot
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>