nvme/auth: take ctrlr lock when getting keys
Currently, the keys are set once, before any of the qpairs are
connected.  However, we'll be changing that and it'll be possible for
users to change the keys at any time.

So, to ensure that the keys are valid while they're in use, they're now
spdk_key_dup()-ed under ctrlr lock.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I60a325c69b6799301144a102b62574b76f342dbe
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/24234
Community-CI: Mellanox Build Bot
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Krzysztof Karas <krzysztof.karas@intel.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Reviewed-by: Ben Walker <ben@nvidia.com>