lib/reduce: merge consecutive IO requests
if chunk_size is 16KB, backing_io_unit is 4KB, when reading
or writing a 16kb of data, it is split into four small IO
requests.
In this patch, if the access area is adjacent, the small IO
is combined into a single IO request, reducing IO latency.

Change-Id: Id8c91854679fcb7f1706a3314dbbe168b4d4f501
Signed-off-by: Yankun Li <845245370@qq.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/23970
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>
Community-CI: Mellanox Build Bot