lib/event: verify if scheduler is not NULL
Fixed issue detected by Coverity.
If the new scheduler could not be initialized, the
spdk_scheduler_set() reverts to the old one, initializing it
again. The issue here is it was done without checking if the
old scheduler (g_scheduler) is not NULL, causing potential
NULL ptr dereference. Now it is fixed.

Change-Id: I7237e529168938e0b8a4cabb7922a69798e2393c
Signed-off-by: Marcin Spiewak <marcin.spiewak@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/24053
Community-CI: Mellanox Build Bot
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>