configure: don't exit when configure idxd on non Intel platforms
In order to better promote DSA to users, it is recommended to change
the error message detected by non Intel CPUs to warning. Users may
compile spdk in docker images with idxd on different platforms,
and then use docker "pull" and "run". But this error affects the
convenient use of DSA. The check here is too strict, and users must
modify the code and configure for different platforms. Although DSA
hardware is not available on non Intel platforms and some lower
versions of Intel CPUs, configuring it is code safe. Because the
order of sending instructions is after checking the DSA hardware.

Signed-off-by: Yin Congmin <congmin.yin@intel.com>