configure: don't exit on non Intel platforms
In order to better promote DSA to users, it is recommended to change
the error message detected by non Intel CPUs to warning. One reason
is that SPDK compilation and execution may not be on the same server.
Another reason is that users may compile spdk when building Docker
images on different platforms. This error is not conducive to directly
compiling idxd into SPDK, then build container images. Since the user
has added “--with-idxd”, we should not block this configure and compile
process. Users need to successfully compile and run it on other
machines or containers on other machines. The check here is too strict.
Although DSA hardware is not available on non Intel platforms and some
lower versions of Intel CPUs, configuring it is code safe. Because the
order of running SPDK and sending instructions is after checking the
DSA hardware.

Change-Id: I378445449827663acad02552abb54cc737f167f0
Signed-off-by: Yin Congmin <congmin.yin@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/24172
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Changpeng Liu <changpeng_liu@hotmail.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>