test/scheduler: fix cpu mask for rpc governor tests
1) When ran through isolate_cores.sh, only specific cores are possible
to use due to cgroup settings. "0x$spdk_main_core" was incorrectly
assigning "0x1" which is core 0. Rather than core 1 as intended.

2) After recent changes gscheduler requires that all SMT cores
are included in CPU mask, so this test wouldn't pass with just one.
Simply use the same spdk_cpumask as in other tests and modify
the remainder of the test to match that.

Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Change-Id: Ieda3e48cb3871a22f45062089daf7aed7b670d1a
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/24303
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>