lib/event: Modify spdk_reactor_set_interrupt_mode() to be called from scheduling reactor
The previous behavior of spdk_reactor_set_interrupt_mode()
returning an error when called from a non-app_thread context
has been changed to return an error when called from a
non-scheduling reactor context. As a result,
spdk_reactor_set_interrupt_mode() must now be executed from a
scheduling core rather than an app thread. Modifications have
been made to address instances where this requirement was not met.

There is no issue when the app thread operates on the scheduling
core, but if it runs on a different core, changing the interrupt
mode will result in an error stating that it can only be modified
on the scheduling core. This could cause scheduling operations
to malfunction. To address this, modifications have been made.

Change-Id: I1f73c84191f5cab877a29405756f106cfb410932
Signed-off-by: syeon.shin <syeon.shin@samsung.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/24324
Community-CI: Mellanox Build Bot
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>