build: Respect -j make option for LTO jobs
We previously used -flto=auto, which attempts to connect to the current
make jobserver to query how many jobs to use. If it can't connect, it
falls back to using the number of CPU cores. GCC itself uses make to run
these LTO jobs.

Since we use make to invoke gcc, and gcc then invokes make, this is a
RECURSIVE make call. In order for the inner-make to find the jobserver
we need to ensure we've done the correct set up for recursive make,
which means prepending the "+" symbol to the line that calls GCC.

We also change from -flto=auto to -flto=jobserver. If it can't find the
job server, it will now produce an error and default to just 1 core.
This is a more sensible behavior than silently failing and defaulting to
all cores, which can lock up build systems.

Change-Id: I4d05a815dbc71f695abb93f208136f741a28118f
Signed-off-by: Ben Walker <ben@nvidia.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/24652
Community-CI: Mellanox Build Bot
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>