lib/bdev: Reset split flag when bounce buffer is used
When read/write IO is sbumitted, we check if IO needs
to be split. That might be true if iovcnt exceeds bdev's
max_sge. However, if memory domain and/or accel sequence
are used, we use a staging buffer and set iovcnt to 1.
In that case no split might be longer required but since
the split flag is still set, we may attempt to split IO,
that leads to some weird results or hit an assert in
debug build.
To fix it, reset split flag once staging buffer is
allocated

Signed-off-by: Alexey Marchuk <alexeymar@nvidia.com>
Change-Id: Ib8a74abbb9e3449ce3e34efd744d3a19623e6815
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/24467
Reviewed-by: Ben Walker <ben@nvidia.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>