blob: add the blob_unref_back_bs_dev helper function.
This is used when a blob doesn't need its current bs_dev anymore. In
current patch we just simply destroy the bs_dev, but in the next commit
we'll extend the logic to the case where there can be multiple refs to
the same bs_dev.

Change-Id: Ie959b681735373400db6d4c56dddf44f6facc316
Signed-off-by: Hadi Moshayedi <hadi@ubicloud.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/24600
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Damiano Cipriani <damiano.cipriani@suse.com>
Reviewed-by: Mike Gerdts <mgerdts@nvidia.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>