test/keyring: ensure session keyring is available
The keyring_linux tests require the session keyring to be available.
Usually, it's created on login and destroyed on logout, so it is always
available, but it's possible to revoke it (keyctl revoke @s), causing
the tests to fail.

So, to avoid relying on system configuration, always run these tests
under a new session.

Fixes: #3492

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: Ie30ef624f6f14a634beb7d8a49b1e3e3e334e92f
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/24774
Reviewed-by: Michal Berger <michal.berger@intel.com>
Reviewed-by: Tomasz Zawadzki <tomasz@tzawadzki.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>