nvme: check process when completing error injected requests
Obviously, requests need to be completed in the process that submitted
them.  However, if one process injects an error to the admin queue while
the other process is polling it, a request with injected errors could be
completed in the wrong process.  To avoid that, we now check the pid of
the process that submitted it before completing it.

Fixes: #3495

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: Ifce5b0cb5dce8adadbae7f3eeb10f019cfd73f0b
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/24790
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Jim Harris <jim.harris@samsung.com>